Remove extraneous sign-in and resolve updates across both files
