+++ app/(account)/sign-in.tsx
@@ -1,5 +1,5 @@
 import { Text, View, StyleSheet } from "react-native";
-import { Button, TextInput, IconButton } from "react-native-paper";
+import { Button, TextInput, IconButton, Portal, Modal, Text as PaperText } from "react-native-paper";
 import {
   MD3LightTheme as DefaultTheme,
   PaperProvider,
@@ -11,6 +11,17 @@ import DarkThemeColors from "@/constants/DarkThemeColors.json";
 import LightThemeColors from "@/constants/LightThemeColors.json";
 import { useColorTheme } from "@/stores/useColorTheme";
 import { Link, router, Stack } from "expo-router";
+import { signInWithEmailAndPassword } from "firebase/auth";
+import { auth } from "@/config/firebaseConfig";
+import { modalStyle } from "@/components/modalStyle";
+import { FirebaseError } from "firebase/app";
+
+export function formatError(error: FirebaseError) {
+  if (error.code === "auth/invalid-credential") {
+    return "Entered credentials do not match an existing user";
+  }
+  return error.message;
+}
 
 export default function SignIn() {
   const screenOptions = {
@@ -27,6 +38,7 @@ export default function SignIn() {
   const [username, setUsername] = useState("");
   const [password, setPassword] = useState("");
   const [passwordVisible, setPasswordVisible] = useState(false);
+  const [error, setError] = useState<unknown | null>(null);
 
   useEffect(() => {
     if (loaded) {
@@ -44,21 +56,21 @@ export default function SignIn() {
       colorTheme === "light" ? LightThemeColors.colors : DarkThemeColors.colors,
   };
 
-  const handleSignIn = () => {
-    console.log(
-      `Attempting to sign in with username: ${username}@ucsd.edu and password: ${password}`
-    );
+  const handleSignIn = async () => {
+    try {
+      const cred = await signInWithEmailAndPassword(auth, username + "@ucsd.edu", password);
+      if (!cred.user.emailVerified) {
+        throw new Error("You must verify your email before using this app! Please check your inbox.");
+      }
 
-    // TODO: Add authentication logic here
-
-    // Go to main page
-    router.push("/(tabs)/main");
+    } catch (fail) {
+      setError(fail);
+    }
   };
 
   return (
-    <>
+    <PaperProvider theme={theme}>
       <Stack.Screen options={screenOptions} />
-      <PaperProvider theme={theme}>
         <View
           style={[styles.container, { backgroundColor: theme.colors.primary }]}
         >
@@ -71,10 +83,12 @@ export default function SignIn() {
               label="Username"
               value={username}
               onChangeText={setUsername}
+              inputMode="text"
               style={styles.input}
               placeholder="Enter your username"
               placeholderTextColor={theme.colors.onBackground}
               right={<TextInput.Affix text="@ucsd.edu" />}
+              textContentType="username"
             />
           </View>
           <View style={styles.inputContainer}>
@@ -85,6 +99,7 @@ export default function SignIn() {
               onChangeText={setPassword}
               style={styles.input}
               placeholder="Enter your password"
+              passwordRules="minlength: 6; required: lower; required: digit;"
               placeholderTextColor={theme.colors.onBackground}
               secureTextEntry={!passwordVisible}
               right={
@@ -116,8 +131,19 @@ export default function SignIn() {
             </Text>
           </View>
         </View>
-      </PaperProvider>
-    </>
+        <Portal>
+        <Modal
+          visible={error !== null}
+          contentContainerStyle={modalStyle.modal}
+        >
+          <PaperText variant="titleMedium">Sign-In Failed</PaperText>
+          <PaperText>
+            {error instanceof FirebaseError ?
+              formatError(error) : `${error}`}
+          </PaperText>
+        </Modal>
+      </Portal>
+    </PaperProvider>
   );
 }
