+++ package.json
@@ -46,14 +46,15 @@
   },
   "devDependencies": {
     "@babel/core": "^7.26.9",
-    "@types/jest": "^29.5.14",
-    "@types/react": "~18.3.18",
-    "@types/react-test-renderer": "^18.3.1",
-    "jest": "^29.7.0",
-    "jest-expo": "~52.0.4",
     "@testing-library/jest-native": "^5.4.3",
     "@testing-library/react-native": "^13.0.1",
+    "@types/jest": "^29.5.14",
+    "@types/react": "~18.3.18",
     "@types/react-native": "^0.72.8",
+    "@types/react-test-renderer": "^18.3.1",
+    "expo-dev-client": "^5.0.12",
+    "jest": "^29.7.0",
+    "jest-expo": "~52.0.4",
     "react-test-renderer": "18.3.1",
     "typescript": "^5.7.3"
   },
